Add tools/prepare_source.sh to distribute source code
GNU GPLv3 requires us to distribute source code when distributing binaries
(see section 6 of GPL). Minified JavaScript can be considered
binary code for the purposes of GPL.

As minified JavaScript is distributed to every user, every user of
Traccar has to be able to download the source code. The best way to do
this is to host the source code archive on the same server the Traccar
web code is served on - i.e. on the Traccar server.

This is convenient especially when the Traccar web code has been modified.

The source code archive is tiny - only around half a megabyte.