MapController: intercepting of reselection events
The user may re-select an existing value from the myReportPeriodField combobox.
In such a case, it may be beneficial to refresh the view.
However, intercepting such an event is tricky as neither 'change' nor
'select' fire in this situation.

Therefore, we apply a trick: we deselect the currently selected value
so that no value is selected.
We also set the text on the combobox manually with
setValueNotFoundText().